+++ cli/tests/test_workspaces.rs
@@ -440,11 +440,11 @@ fn test_workspaces_add_workspace_in_current_workspace() {
 
     // Workspace created despite warning
     let output = main_dir.run_jj(["workspace", "list"]);
-    insta::assert_snapshot!(output, @r"
-    default: rlvkpnrz af2d0cd5 (no description set)
+    insta::assert_snapshot!(output, @r###"
+    default: rlvkpnrz 504e3d8c (empty) (no description set)
     secondary: pmmvwywv 058f604d (empty) (no description set)
     [EOF]
-    ");
+    "###);
 
     // Use explicit path instead (no warning)
     let output = main_dir.run_jj(["workspace", "add", "./third"]);
@@ -459,22 +459,18 @@ fn test_workspaces_add_workspace_in_current_workspace() {
 
     // Both workspaces created
     let output = main_dir.run_jj(["workspace", "list"]);
-    insta::assert_snapshot!(output, @r"
-    default: rlvkpnrz b2288847 (no description set)
+    insta::assert_snapshot!(output, @r###"
+    default: rlvkpnrz 504e3d8c (empty) (no description set)
     secondary: pmmvwywv 058f604d (empty) (no description set)
     third: zxsnswpr 1c1effec (empty) (no description set)
     [EOF]
-    ");
+    "###);
 
-    // Can see files from the other workspaces in main workspace, since they are
-    // child directories and will therefore be snapshotted
     let output = main_dir.run_jj(["file", "list"]);
-    insta::assert_snapshot!(output.normalize_backslash(), @r"
+    insta::assert_snapshot!(output.normalize_backslash(), @r###"
     file
-    secondary/file
-    third/file
     [EOF]
-    ");
+    "###);
 }
 
 /// Test making changes to the working copy in a workspace as it gets rewritten