+++ lib/src/local_working_copy.rs
@@ -1371,6 +1371,20 @@ impl FileSnapshotter<'_> {
 
         if file_type.is_dir() {
             let file_states = file_states.prefixed_at(dir, name);
+            // If a submodule was added in commit C, and a user decides to run
+            // `jj new <something before C>` from after C, then the submodule
+            // files stick around but it is no longer seen as a submodule.
+            // We need to ensure that it is not tracked as if it was added to
+            // the main repo.
+            // See https://github.com/jj-vcs/jj/issues/4349.
+            // To solve this, we ignore all nested repos entirely.
+            let disk_dir = entry.path();
+            for &name in RESERVED_DIR_NAMES {
+                if disk_dir.join(name).symlink_metadata().is_ok() {
+                    return Ok(None);
+                }
+            }
+
             if git_ignore.matches(&path.to_internal_dir_string())
                 && self.force_tracking_matcher.visit(&path).is_nothing()
             {
@@ -1383,7 +1397,7 @@ impl FileSnapshotter<'_> {
             } else if !self.matcher.visit(&path).is_nothing() {
                 let directory_to_visit = DirectoryToVisit {
                     dir: path,
-                    disk_dir: entry.path(),
+                    disk_dir,
                     git_ignore: git_ignore.clone(),
                     file_states,
                 };